+++ cmd/influxd/run/config.go
@@ -31,12 +31,8 @@ import (
 )
 
 const (
-	// DefaultBindAddress is the default address for raft, cluster, snapshot, etc..
+	// DefaultBindAddress is the default address for various RPC services.
 	DefaultBindAddress = ":8088"
-
-	// DefaultHostname is the default hostname used if we are unable to determine
-	// the hostname from the system
-	DefaultHostname = "localhost"
 )
 
 // Config represents the configuration format for the influxd binary.
@@ -63,12 +59,6 @@ type Config struct {
 
 	// BindAddress is the address that all TCP services use (Raft, Snapshot, Cluster, etc.)
 	BindAddress string `toml:"bind-address"`
-
-	// Hostname is the hostname portion to use when registering local
-	// addresses.  This hostname must be resolvable from other nodes.
-	Hostname string `toml:"hostname"`
-
-	Join string `toml:"join"`
 }
 
 // NewConfig returns an instance of Config with reasonable defaults.