+++ services/admin/config.go
@@ -11,7 +11,7 @@ type Config struct {
 	BindAddress      string `toml:"bind-address"`
 	HTTPSEnabled     bool   `toml:"https-enabled"`
 	HTTPSCertificate string `toml:"https-certificate"`
-	Version          string
+	Version          string `toml:"-"`
 }
 
 // NewConfig returns an instance of Config with defaults.